[judge] Fix login, begin, vote UX flow
Clicking "Done" on the begin page didn't render the `vote` template 
because on POST to the judge:begin URL, we would redirect to the 
judge:vote URL but the `vote` view would not be executed (likely because 
we were redirecting from /judge/vote to /judge/vote).

To fix this, we changed the begin view to allow GET (and POST) requests, 
and we render the `begin` template in the begin view.