[judge] Do not compress timer JS
This commit is a hotfix for 
`compressor.exceptions.OfflineGenerationError: You have offline 
compression enabled but key "..." is missing from offline manifest. You 
may need to run "python manage.py compress"`. This issue occurrs because 
each render of the template returns a different `script` tag since the 
random suffix is dynamic. 

This issue has three solutions: 
1. do not compress the timer `script` tag
2. switch from offline compression to in-request compression
3. (potentially) setting COMPRESS_OFFLINE_CONTEXT

where the list is ordered by how easy each solution is to implement. We 
choose option 1 over option 2 because we would rather use offline 
compression (which is faster) for everywhere except the timer than use 
in-request compression (which is slower) for the entire site.